调试上传目录

+++ src/xapi/xapi.h
@@ -29,7 +29,7 @@
 
 typedef struct __xfs_item {
     int type;
-    int path_len;
+    uint64_t path_len;
     const char *path;
     uint64_t size;
 }*__xfs_item_ptr;
